requirements.txt
flask
python-dotenv
flask-sqlalchemy
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
